ignore curl dependabot updates + update curl to 8.8.0 + update remaining git submodules + update sentry-native to 0.7.5

+++ bmedll/Updater.cpp
@@ -136,7 +136,7 @@ namespace Updater {
 
             if (res != CURLE_OK)
             {
-                spdlog::error("Update download error, curl code is not CURLE_OK: {}", res);
+                spdlog::error("Update download error, curl code is not CURLE_OK: {}", (int)res);
                 return L"";
             }
             if (response_code != 200)
